@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.3",
+        "php": "^7.4",
         "ext-json": "*",
         "beyondcode/laravel-self-diagnosis": "^1.3",
         "fideloper/proxy": "^4.0",
