@@ -25,12 +25,11 @@ class RemoveGroupMembers
         // Remove users from the input array, that are not a member of the group.
         $currentMembers = collect($group->users()->pluck('id'));
         $newMembers = collect($args['members'])
-            ->filter(function ($item) use ($currentMembers) {
-                return $currentMembers->contains($item);
-            })
+            ->filter(fn ($item) => $currentMembers->contains($item))
             ->toArray();
 
         $group->users()->detach($args['members']);
+
         return $group;
     }
 }
